+++ crypto/src/basic/anemoi_jive/salts.rs
@@ -1,0 +1,106 @@
+use noah_algebra::bls12_381::BLSScalar;
+use noah_algebra::new_bls12_381;
+
+/// The salts for the Anemoi-Jive CRH over BLS12-381.
+/// The salts are generated with the following code.
+/// ```ignore
+/// pub static PI_2: &str = "44288109756659334461284756482337867831652712\
+/// 01909145648566923460348610454326648213393607260249141273";
+/// pub static PI_3: &str = "72458700660631558817488152092096282925409171\
+/// 53643678925903600113305305488204665213841469519415116094";
+///
+/// // Generate the salts
+/// pub fn generate_salts<F: PrimeField>(generator: &F, alpha: u32, num: usize) -> Vec<F> {
+///     let pi_2_biguint = BigUint::from_str(PI_2).unwrap();
+///     let pi_3_biguint = BigUint::from_str(PI_3).unwrap();
+///
+///     let prime_biguint: BigUint = F::MODULUS.into();
+///
+///     let pi_f_2 = F::from(pi_2_biguint % &prime_biguint);
+///     let pi_f_3 = F::from(pi_3_biguint % &prime_biguint);
+///
+///     let mut salts = Vec::<F>::new();
+///
+///     let mut pi_2_r = F::one();
+///     for r in 0..num {
+///         let mut pi_3_i = F::one();
+///
+///         let pow_alpha: F = (pi_2_r + pi_3_i).pow(&[alpha as u64]);
+///         salts.push(*generator * pi_2_r.square() + pow_alpha);
+///         pi_3_i *= pi_f_3;
+///
+///         pi_2_r *= pi_f_2;
+///     }
+///
+///     salts
+/// }
+///
+/// let salts = generate_salts::<ark_bls12_381::Fr>(&ark_bls12_381::Fr::from(7u64), 5u32, 64);
+///
+/// ```
+pub const ANEMOI_JIVE_381_SALTS: [BLSScalar; 64] = [
+    new_bls12_381!("6406215194479240286762731634835344236141886914605144794931128113894074089386"),
+    new_bls12_381!("25560080366671527635336967422834298208909930660967190727048965370381122828324"),
+    new_bls12_381!("50658439267116975037933099803088424427085069111100904739841927317887955508403"),
+    new_bls12_381!("49202867374801348233366257521250211085729311964982431211168378339150325266596"),
+    new_bls12_381!("39512762860131380300143104002628333366763410802467663674311566818148974565133"),
+    new_bls12_381!("11876717497669986095596439982191006521736331858730684113164765237413845473944"),
+    new_bls12_381!("35905493000702460707851319694883808822917434543549898626738506946150871324421"),
+    new_bls12_381!("27186483893192137888367056646633699857046736175736979525890113166774163514623"),
+    new_bls12_381!("17106901080657171059221588430697500200139148930560720708527494771434207741325"),
+    new_bls12_381!("2947247105774581031713063065251521320329902750845167986294561654101405196807"),
+    new_bls12_381!("51592079572413191872853358671431032085743619585916994428905417867123890278387"),
+    new_bls12_381!("24816805674044088753437614576850034920688912721097761569100178771751601655270"),
+    new_bls12_381!("18737304489188221477320550535667941529509106215289172413722189957955490476880"),
+    new_bls12_381!("13208403487417056542747186342055085956605253382587954734743695712810536004854"),
+    new_bls12_381!("22174026916692900500018984239981888364960054940583861997478529762257435881649"),
+    new_bls12_381!("14277223701616846144752813397523985951727042947802894661372383492716148000958"),
+    new_bls12_381!("38537729790878817151832254066940670734709071874829822789012367620206294295697"),
+    new_bls12_381!("46429159945618798562266749928924897987650987233446873772298366545774012057031"),
+    new_bls12_381!("49573185258004972705338141619067910819644923361360150569716410656138196617350"),
+    new_bls12_381!("8808549100351786545270432628901850098657939587763329761465451280385563386019"),
+    new_bls12_381!("47545884453564576395283966691168696006505461953304970106904616669538449617563"),
+    new_bls12_381!("21482870545546346116520674895805037239909053292564410348875474566920600909715"),
+    new_bls12_381!("33186690764244511345193651820558716649119982023617916307826540102554516472139"),
+    new_bls12_381!("30350340353888399331126590000673687431830313888658096223457706922661233825688"),
+    new_bls12_381!("3942545902701090087705826329290868532706337657003733728810633366482846729506"),
+    new_bls12_381!("2082510624392818698732847366299654635857775777539656452414916165122403066195"),
+    new_bls12_381!("51649334795838988372125488154003249393791640708325093419815644887794813309008"),
+    new_bls12_381!("34887858772413423169834803418440531108993150189891335581819239064760286101455"),
+    new_bls12_381!("48936583161825646569724419127726438460187933985953245447187452101796146735887"),
+    new_bls12_381!("32642730579209546249308615043133770609665519724015393385939652278500890099536"),
+    new_bls12_381!("13613672512975190978853575030239289621055291996871872964870881328315332173445"),
+    new_bls12_381!("45857815948615754399416976418518399475658898258642746840125919814552021343539"),
+    new_bls12_381!("27930434012335308503985612474694399710070769233460431210898055589039696617174"),
+    new_bls12_381!("16082891484083197962298420031939300179737546889483404481270668239938925730519"),
+    new_bls12_381!("32431358382254799316517534547310733442268246743646647830932614874662926374169"),
+    new_bls12_381!("12001798800269343419894608733509428807939395673673182558768306347518155928368"),
+    new_bls12_381!("19091998626622759830887219341787346217536406773488965115213063023260188150196"),
+    new_bls12_381!("4061365125589423535263386040342356261325929153772201912893015669652043473651"),
+    new_bls12_381!("42723069573372575409439879325916774325247419381280869222846543303609926685073"),
+    new_bls12_381!("33411013226341788489531212408025099328807232565519108168173254434046156358114"),
+    new_bls12_381!("22285215956726326717555645346885663035321379032235291157141463562680247858723"),
+    new_bls12_381!("19100716338800473714612639247494110760560181407251931992121147800828314529681"),
+    new_bls12_381!("37908871663584165335205509051952483202836179525323491895871954224141938867990"),
+    new_bls12_381!("4271765406273011823704321421282739325904910283385044968295660170595223240215"),
+    new_bls12_381!("50876843920929781237572991363935625089207791214968266941968601163581536883058"),
+    new_bls12_381!("1177156943948505448921151255413373199132597755232043043050169936690894579252"),
+    new_bls12_381!("51587349655410657129411563705348793062261635176459590250788447130472536465786"),
+    new_bls12_381!("9219550859140898932258056590169454588152106345986798523391093622183609260757"),
+    new_bls12_381!("50031287409312337488777320835097641243752192579702150358835452882985744695259"),
+    new_bls12_381!("32873249177724826442701901791509764378077343046877699374690979165865248847923"),
+    new_bls12_381!("16761444014354370430823854639861333789647902397513323110799773505932848409160"),
+    new_bls12_381!("40185192398708244083559562678736292297196925796969428219679037536661994618983"),
+    new_bls12_381!("27811697745704948617150508783769544237268384518115636221517802995039085883956"),
+    new_bls12_381!("11836823182912063961934875161323318469855118006229360566507084183152286812991"),
+    new_bls12_381!("52119247984727570142503947637528636360850027343934935965194199382704200721131"),
+    new_bls12_381!("12061754451123296293631777903065177515680659949636535501781111345784880536267"),
+    new_bls12_381!("43479045165837144162744599062941533175485374356835125667928524388351515117313"),
+    new_bls12_381!("23103429557191913769940047466046366751339316644963955862724231703064129035334"),
+    new_bls12_381!("254369426964247362468237865478019065502192043444834575103908764163406492126"),
+    new_bls12_381!("48071597869367984847725917733368171763615703896109771496842923848158767414061"),
+    new_bls12_381!("15055256968173980806767925613812067038912295204458123756113946263570286560712"),
+    new_bls12_381!("45304830436629094174821131905656972358424981966498279424220949086434482388311"),
+    new_bls12_381!("20011523849368630227591426012028344136556818717871466977616177645897908380509"),
+    new_bls12_381!("20913680202770605374619314867123384851639737147442790410915664878667151333977"),
+];
